server/requirements.txt
alembic==1.12.1
APScheduler==3.10.4
backports.zoneinfo==0.2.1
certifi==2024.2.2
chardet==3.0.4
click==7.1.2
cymysql==0.9.18
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.4
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.4.4
gevent==21.12.0
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==2.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.4
MarkupSafe==1.1.1
monotonic==1.6
pycryptodome==3.10.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2024.1
qiniu==7.2.10
redis==3.5.3
requests==2.23.0
six==1.16.0
SQLAlchemy==1.3.24
typing_extensions==4.7.1
tzlocal==5.1
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==3.0.1
zipp==3.15.0
zope.event==5.0
zope.interface==6.3
